Add deputes_id.json
A dictionnary of id linking to the name and surname of a deputy.

It's a cleaning of http://www.nosdeputes.fr/deputes/enmandat/json that
is kind of weird and uneasy to use. Cleaning done like this:

import json

f = open('deputes.json', 'r')
deputes = json.load(f)

new_deputes = {}

for depute in deputes["deputes"]:
    new_deputes[depute["depute"]["id_an"]] = {}
    new_deputes[depute["depute"]["id_an"]]["prenom"] = depute["depute"]["nom"]
    new_deputes[depute["depute"]["id_an"]]["nom"] = depute["depute"]["nom_de_famille"]

print json.dumps(new_deputes)